Use a *pgxpool.Pool instead of a *pgx.Conn (#5)
* Use a *pgxpool.Pool instead of a *pgx.Conn

The motivation for this is tests that are testing explicitly for concurrency, or require multiple queries in parallel. Using a single conn in such cases will cause problems, as conns are not concurrency-safe.

pgxpool.Pool is concurrency-safe though, so we use that with a sensible number of max conns.

In the future, we might want to make this configurable.

* Fix tests